requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
plotly>=5.0.0
python-docx>=0.8.11
haversine>=2.8.0
supabase==1.2.0
gotrue==1.3.0
httpx==0.24.1
